requirements.txt
facebook-sdk
openai
aiohttp
beautifulsoup4
requests
python-dotenv
